HTTPS: to remove CORS errors in Express run
WebSocket handshake errors gone in the WebQuizProject express app. The current configuration for building the static /dist files for the viewer are pointing to https://localhost/ohif (no port).
This is how Nginx docker containers are serving the files and it plays nicely with the WebQuiz app which loads ohif in the iframe at https://localhost/ohif .
AS LONG AS YARN START:HTTPS is not run after the YARN BUILD command, the port is not injected.
yarn start:https uses the devServer configuration which automatically injects port 3000 into the origin.